requirements.txt
Flask
supabase-py
gunicorn # Gunicorn is a WSGI HTTP server for Unix, required by Render for Python web services.